refactor coral wrist to be separate from coral intake
