+++ src/main/java/frc/robot/CommandFactory.java
@@ -1,5 +1,7 @@
 package frc.robot;
 
+import edu.wpi.first.wpilibj2.command.Command;
+
 public class CommandFactory {
 
     public CommandFactory() {
