+++ src/main/java/frc/robot/RobotContainer.java
@@ -11,6 +11,7 @@ import frc.robot.Constants.DriveControlConstants;
 import frc.robot.Constants.SetpointConstants;
 import frc.robot.Constants.SpeedConstants;
 import frc.robot.subsystems.coralIntake.CoralIntakeSubsystem;
+import frc.robot.subsystems.coralWrist.CoralWristSubsystem;
 import frc.robot.subsystems.drive.DriveSubsystem;
 import frc.robot.subsystems.gyro.Gyro;
 import frc.robot.vision.*;
@@ -21,6 +22,7 @@ public class RobotContainer {
   private Gyro gyro = subsystemFactory.buildGyro();
   private DriveSubsystem drive = subsystemFactory.buildDriveSubsystem(gyro);
   private final CoralIntakeSubsystem coralIntake = subsystemFactory.buildCoralIntake();
+  private final CoralWristSubsystem coralWrist = subsystemFactory.buildCoralWrist();
   // this is public because we need to run the visionPoseEstimator periodic from
   // Robot
   public VisionPoseEstimator visionPoseEstimator = new VisionPoseEstimator(drive);
@@ -53,7 +55,8 @@ public class RobotContainer {
                     DriveControlConstants.DRIVE_DEADBAND)),
                 DriveControlConstants.FIELD_ORIENTED_DRIVE),
             drive).withName("drive default"));
-    coralIntake.setDefaultCommand(coralIntake.setSpeeds());
+    coralIntake.setDefaultCommand(coralIntake.setRollerSpeed(0));
+    coralWrist.setDefaultCommand(coralWrist.setWristSpeed(0));
   }
 
   private void configureButtonBindingsDriver() {
@@ -62,8 +65,8 @@ public class RobotContainer {
   }
 
   private void configureButtonBindingsOperator() {
-    operatorController.y().whileTrue(coralIntake.setWristSpeed(SpeedConstants.WRIST_MAX_SPEED));
-    operatorController.a().whileTrue(coralIntake.setWristSpeed(-SpeedConstants.WRIST_MAX_SPEED));
+    operatorController.y().whileTrue(coralWrist.setWristSpeed(SpeedConstants.WRIST_MAX_SPEED));
+    operatorController.a().whileTrue(coralWrist.setWristSpeed(-SpeedConstants.WRIST_MAX_SPEED));
     operatorController.x().whileTrue(coralIntake.setRollerSpeed(SpeedConstants.CORAL_INTAKE_SPEED));
     operatorController.b().whileTrue(coralIntake.setRollerSpeed(SpeedConstants.CORAL_OUTTAKE_SPEED));
   }
