+++ src/main/java/frc/robot/subsystems/coralIntake/CoralIntakePlacebo.java
@@ -1,27 +1,22 @@
 package frc.robot.subsystems.coralIntake;
 
-import edu.wpi.first.units.measure.Angle;
-
 public class CoralIntakePlacebo implements CoralIntakeIO {
-   public void setRollerSpeed(double speed) {
-       speed = 0;
-   }
 
-   public void setWristSpeed(double speed) {}
+    @Override
+    public void setRollerSpeed(double speed) {
+    }
 
-   public void setGravityCompensation() {}
+    @Override
+    public double getVelocity() {
+        return 0.0;
+    }
 
-   public double getVelocity() {
-       return 0.0;
-   }
+    @Override
+    public double getCurrent() {
+        return 0.0;
+    }
 
-   public double getCurrent() {
-       return 0.0;
-   }
-
-   public void goToSetpoint(Angle angle) {
-   }
-
-   public void updateStates(CoralIntakeIOStates states) {
-   }
+    @Override
+    public void updateStates(CoralIntakeIOStates states) {
+    }
 }
