+++ src/main/java/frc/robot/subsystems/coralIntake/CoralIntakeSubsystem.java
@@ -1,9 +1,7 @@
 package frc.robot.subsystems.coralIntake;
 
-import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.subsystems.coralIntake.CoralIntakeIO.CoralIntakeIOStates;
 
@@ -17,27 +15,6 @@ public class CoralIntakeSubsystem extends SubsystemBase {
 
     public Command setRollerSpeed(double speed) {
         return this.run(() -> io.setRollerSpeed(speed));
-    }
-
-    public Command goToSetpoint(Angle angle) {
-        return this.run(() -> {
-            io.goToSetpoint(angle);
-            io.setRollerSpeed(0);
-        });
-    }
-
-    public Command setWristSpeed(double speed) {
-        return this.run(() -> {
-            io.setWristSpeed(speed);
-            io.setRollerSpeed(0);
-        });
-    }
-
-    public Command setSpeeds() {
-        return this.run(() -> {
-            io.setWristSpeed(0);
-            io.setRollerSpeed(0);
-        });
     }
 
     @Override
