+++ src/main/java/frc/robot/subsystems/coralWrist/CoralWristPlacebo.java
@@ -1,0 +1,18 @@
+package frc.robot.subsystems.coralWrist;
+
+import edu.wpi.first.units.measure.Angle;
+
+public class CoralWristPlacebo implements CoralWristIO {
+
+    @Override
+    public void goToSetpoint(Angle angle) {
+    }
+
+    @Override
+    public void setWristSpeed(double speed) {
+    }
+
+    @Override
+    public void updateStates(CoralWristIOStates states) {
+    }
+}
